corregi un problema ahi al querer ingresar un profe en una escuela que no se encontrara

+++ Matricula/Matricula/aranceles.cpp
@@ -1,11 +1,11 @@
 #include "aranceles.h"
 
-float Aranceles::cobroMatricula(Curso *C, Estudiante *E)
+double Aranceles::cobroMatricula(Curso *C, Estudiante *E)
 {
-	float resultado = 0;
-	resultado = (float) C->getCantidadCreditos() * costoCredito;
+	double resultado = 0;
+	resultado =  C->getCantidadCreditos() * costoCredito;
 	resultado = resultado + cargosAdministrativos;
 	if (E->getNacionalidad())
-		resultado = resultado + (resultado * recargoExtranjeros);
+		resultado =  resultado + (resultado * recargoExtranjeros);
 	return resultado;
 }
